Recover `et-user` from missing config directory
`et-user` alread had the means to recover from a missing `user.json` file. If
however, for what ever reason, the whole emcomm-tools config folder is missing,
it spits out all sorts of errors. While the errors would be just a UX issue, it
is unable to recover and the system would remain in a broken state.

Recreating the missing folder when missing, mitigates this issue.